Add files via upload
The 'Rules' file will work on the dataset 'wrokouts.csv' to filter out user specific workouts depending upon thier intensities and diffculties. This is being done in the first two cells of code in the ipython file. The third code cell then further works on these filtered workouts to populate the filtered workouts by assigning priority sccoring to the workouts specific to user's goals and area of focus.
The hard coded input is as follows: 
In the second code cell look for a function 'get_user_input' it is entering a user id, user age and activity level as 'string', integer and 'string' data formats, respectively.
In the third code cell the hard coded input is lying in plain sight in the last few lines. There is a redundancy of user_id declaration ignor it for now.