Add files via upload
Don't mind this file this just creates a backbone for the actual workout plan for users with different goals. When executed it'll print a 30,60 or 90 days workout plan with weekly iterations of the same focus areas and rest days. i will be populating this backbone plan with the workouts of user specific intensity and difficulty in another script this is just a reference script ....kind of.